Move rename encoding tests into existing client_test.go files
Delete the separate rename_encoding_test.go files and fold the
captureTransport helper and test functions into the existing
client_test.go in both file/ and directory/ packages, matching
the repo convention of one test file per package.
