+++ sdk/storage/azdatalake/file/client_test.go
@@ -6490,6 +6490,43 @@ func (s *UnrecordedTestSuite) TestFileGetSetTagsFileIdentitySas() {
 	_require.Equal(tags["tagKey1"], tagMap["tagKey1"])
 }
 
+type captureTransport struct {
+	req *http.Request
+}
+
+func (c *captureTransport) Do(req *http.Request) (*http.Response, error) {
+	c.req = req
+	return &http.Response{
+		Request:    req,
+		Status:     "Created",
+		StatusCode: http.StatusCreated,
+		Header:     http.Header{},
+		Body:       http.NoBody,
+	}, nil
+}
+
+func TestFileRenameEncodesSourcePath(t *testing.T) {
+	_require := require.New(t)
+	ct := &captureTransport{}
+
+	srcURL := "https://fake.dfs.core.windows.net/myfs/dir1/l%C3%B6r%20006.jpg"
+	fClient, err := file.NewClientWithNoCredential(srcURL, &file.ClientOptions{
+		ClientOptions: policy.ClientOptions{Transport: ct},
+	})
+	_require.NoError(err)
+
+	_, err = fClient.Rename(context.Background(), "dir1/renamed.jpg", nil)
+	_require.NoError(err)
+	_require.NotNil(ct.req)
+
+	renameSourceVals := ct.req.Header["x-ms-rename-source"] //nolint:staticcheck // SA1008: the generated client stores this header under a non-canonical key, so it must be read with the same raw key.
+	_require.NotEmpty(renameSourceVals)
+	renameSource := renameSourceVals[0]
+	_require.Contains(renameSource, "l%C3%B6r%20006.jpg")
+	_require.NotContains(renameSource, " ")
+	_require.NotContains(renameSource, "ö")
+}
+
 func (s *UnrecordedTestSuite) TestFileGetSetTagsFileSystemIdentitySas() {
 	// Datalake tags is currently in public preview and not GA yet, skipping this test for now.
 	s.T().Skip("Datalake tags is in public preview and not GA yet")
